Promotions Thingy/requirements.txt
streamlit
firebase-admin
pandas
google-generativeai
tabulate
plotly